creado create trip

+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { TokenService } from './services/token.service';
-import { Router } from '@angular/router';
+import { NavigationEnd, Router } from '@angular/router';
 
 @Component({
   selector: 'app-root',
@@ -11,17 +11,25 @@ import { Router } from '@angular/router';
 export class AppComponent implements OnInit {
 
   title = 'FrontAng';
+  private publicRoutes = ['/login', '/register'];
   constructor(private tokenService: TokenService,private router: Router) {}
 
   ngOnInit() {
-    this.tokenService.checkTokenValidity().subscribe(isValid => {
-      if (!isValid) {
-        console.log('Token is invalid or expired');
-        this.router.navigate(['/login']);
-        this.tokenService.removeToken();
-      } else {
-        console.log('Token is valid');
-        // Continúa con la lógica de la aplicación
+    this.router.events.subscribe(event => {
+      if (event instanceof NavigationEnd) {
+        if (this.publicRoutes.includes(event.url)) {
+          return;
+        }
+        // Validar el token solo en rutas privadas
+        this.tokenService.checkTokenValidity().subscribe(isValid => {
+          if (!isValid) {
+            console.log('Token is invalid or expired');
+            this.tokenService.removeToken();
+            this.router.navigate(['/login']);
+          } else {
+            console.log('Token is valid');
+          }
+        });
       }
     });
   }
